3. Int to long

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        SalesManager stats = new SalesManager(new int[]{10, 20, 15, -10, 5});
+        SalesManager stats = new SalesManager(new long[]{10, 20, 15, -10, 5});
         System.out.println("Максимальная продажа: " + stats.max());
     }
 }
